fix(ci): drop eval from contrast gate and share dist server
Sonar flagged eval as a security hotspot and duplicated static
preview helpers. Reuse one server helper and keep dark CTA CSS
scoped to bg-gray-950.
